after review refactor code
