+++ task9mvvm/src/main/java/by/itacademy/training/task9mvvm/model/domain/ForecastDay.kt
@@ -1,3 +1,3 @@
-package by.itacademy.training.task9mvvm.model.entity
+package by.itacademy.training.task9mvvm.model.domain
 
 data class ForecastDay(val list: List<HourTemperature>)
