+++ task9mvvm/src/main/java/by/itacademy/training/task9mvvm/model/domain/TemperatureUnit.kt
@@ -1,4 +1,4 @@
-package by.itacademy.training.task9mvvm.util
+package by.itacademy.training.task9mvvm.model.domain
 
 enum class TemperatureUnit {
     FAHRENHEIT,
