+++ task9mvvm/src/main/java/by/itacademy/training/task9mvvm/model/domain/WeatherReport.kt
@@ -1,4 +1,4 @@
-package by.itacademy.training.task9mvvm.model.entity
+package by.itacademy.training.task9mvvm.model.domain
 
 data class WeatherReport(
     val currentTemperature: CurrentTemperature,
