+++ task9mvvm/src/main/java/by/itacademy/training/task9mvvm/model/repository/WeatherForecastRepositoryImpl.kt
@@ -1,14 +1,13 @@
 package by.itacademy.training.task9mvvm.model.repository
 
 import by.itacademy.training.task9mvvm.model.api.WeatherApi
-import by.itacademy.training.task9mvvm.util.DTOMapper
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 class WeatherForecastRepositoryImpl @Inject constructor(
     private val api: WeatherApi,
-    private val dtoMapper: DTOMapper
+    private val dtoMapper: WeatherReportMapper
 ) : WeatherForecastRepository {
 
     override suspend fun getWeatherForecastForDay(city: String) = withContext(Dispatchers.IO) {
