+++ task9mvvm/src/main/java/by/itacademy/training/task9mvvm/model/repository/WeatherReportMapper.kt
@@ -1,17 +1,18 @@
-package by.itacademy.training.task9mvvm.util
+package by.itacademy.training.task9mvvm.model.repository
 
+import by.itacademy.training.task9mvvm.model.domain.Condition
+import by.itacademy.training.task9mvvm.model.domain.CurrentTemperature
+import by.itacademy.training.task9mvvm.model.domain.ForecastDay
+import by.itacademy.training.task9mvvm.model.domain.HourTemperature
+import by.itacademy.training.task9mvvm.model.domain.Location
+import by.itacademy.training.task9mvvm.model.domain.WeatherReport
 import by.itacademy.training.task9mvvm.model.dto.api.ApiResponse
-import by.itacademy.training.task9mvvm.model.entity.Condition
-import by.itacademy.training.task9mvvm.model.entity.CurrentTemperature
-import by.itacademy.training.task9mvvm.model.entity.ForecastDay
-import by.itacademy.training.task9mvvm.model.entity.HourTemperature
-import by.itacademy.training.task9mvvm.model.entity.Location
-import by.itacademy.training.task9mvvm.model.entity.WeatherReport
+import by.itacademy.training.task9mvvm.util.Formatter
 import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
-class DTOMapper @Inject constructor(private val formatter: MetricFormatter) :
+class WeatherReportMapper @Inject constructor(private val formatter: Formatter) :
     (ApiResponse) -> WeatherReport {
 
     override fun invoke(resposne: ApiResponse): WeatherReport {
