+++ task9mvvm/src/main/java/by/itacademy/training/task9mvvm/util/SupportSharedPreference.kt
@@ -1,6 +1,7 @@
 package by.itacademy.training.task9mvvm.util
 
 import android.content.SharedPreferences
+import by.itacademy.training.task9mvvm.model.domain.TemperatureUnit
 
 interface SupportSharedPreference {
